[jk] Add pipelines folder with metadata file visible (#1237)
* [jk] Show pipelines folder with all pipelines and config files

* [jk] Fix error

* [jk] Fix context menu issue in file browser

* [jk] Disable unsupported file types

* [jk] Remove unused context menu item

* [jk] Update integrations readme links

+++ mage_ai/frontend/components/ContextMenu/index.tsx
@@ -99,13 +99,9 @@ function ContextMenu({
   }, []);
 
   const handleClick = useCallback((e) => {
-    if (visible) {
-      e.preventDefault();
-      setVisible(false);
-    }
+    setVisible(false);
   }, [
     setVisible,
-    visible,
   ]);
 
   const handleContextMenu = useCallback((e: MouseEvent) => {
@@ -131,6 +127,9 @@ function ContextMenu({
         label: () => 'Delete',
         onClick: () => {
           const { block } = contextItem.data;
+          if (!block) {
+            return;
+          }
           if (block.type === BlockTypeEnum.CHART) {
             deleteWidget(block);
           } else {
@@ -140,24 +139,6 @@ function ContextMenu({
         uuid: 'delete_block_file',
       },
     ],
-    // [FileContextEnum.PIPELINE]: [
-    //   {
-    //     label: () => 'Duplicate',
-    //     onClick: () => createPipeline({
-    //       pipeline: {
-    //         clone_pipeline_uuid: contextItem.data.name,
-    //         name: `${contextItem.data.name}_copy`,
-    //       },
-    //     }),
-    //     uuid: 'duplicate_pipeline',
-    //   },
-    //   {
-    //     disabled: numPipelines <= 1,
-    //     label: () => 'Delete',
-    //     onClick: () => deletePipeline(contextItem.data.name),
-    //     uuid: 'delete_pipeline',
-    //   },
-    // ],
   };
 
   return (
